add INSERT and DROP for Lesson entity + ui. Fix BUGS for Course Entity

+++ src/main/java/io/fnska/blog/site/controller/CourseController.java
@@ -3,7 +3,6 @@ package io.fnska.blog.site.controller;
 import io.fnska.blog.site.domain.Course;
 import io.fnska.blog.site.service.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -19,30 +18,44 @@ public class CourseController {
     @Autowired
     private CourseService courseService;
 
-    @RequestMapping(value = "/courses", method = RequestMethod.GET)
-    public String getAllCourses(Model model) {
-        model.addAttribute("th_courses", courseService.getAllCourses());
-        return "complex/courses";
+    @RequestMapping(value = "user/{username}/courses", method = RequestMethod.GET)
+    public String getAllCoursesByUsername(@PathVariable("username") String username,
+                                          Principal principal, Model model) {
+        if (username.equalsIgnoreCase(principal.getName())) {
+            model.addAttribute("th_courses", courseService.getAllCoursesByUsername(principal));
+            return "complex/courses";
+        }
+        return "error/403";
     }
 
-    @RequestMapping(value = "/courses/{courseYear}", method = RequestMethod.GET)
-    public String getCourse(@PathVariable String courseYear, Model model) {
-        model.addAttribute("th_course", courseService.getCourse(courseYear));
-        return "single/course";
+    @RequestMapping(value = "user/{username}/courses/{courseYear}", method = RequestMethod.GET)
+    public String getCourse(@PathVariable("username") String username,
+                            @PathVariable String courseYear,
+                            Principal principal, Model model) {
+        if (username.equalsIgnoreCase(principal.getName())) {
+            model.addAttribute("th_course", courseService.getCourseByYearAndUsername(principal, courseYear));
+            return "single/course";
+        }
+        return "error/403";
     }
 
-    @PreAuthorize("hasAuthority('ADMIN')")
-    @RequestMapping(value = "/admin/create-course", method = RequestMethod.POST)
-    public String addCourse(@ModelAttribute Course course, Principal principal) {
-        courseService.addCourse(course, principal);
-        return "redirect:/admin/dashboard";
+    @RequestMapping(value = "user/{username}/dashboard/create-course", method = RequestMethod.POST)
+    public String addCourse(@PathVariable("username") String username,
+                            @ModelAttribute Course course, Principal principal) {
+        if (username.equalsIgnoreCase(principal.getName())) {
+            courseService.addCourse(course, principal);
+            return "redirect:/user/{username}/dashboard";
+        }
+        return "error/403";
     }
 
-    @PreAuthorize("hasAuthority('ADMIN')")
-    @RequestMapping(value = "/admin/delete-course", method = RequestMethod.POST)
-    public String deleteCourse(@ModelAttribute Course course, Principal principal) {
-        courseService.deleteCourse(course.getYear(), principal);
-        return "redirect:/admin/dashboard";
+    @RequestMapping(value = "user/{username}/dashboard/delete-course", method = RequestMethod.POST)
+    public String deleteCourse(@PathVariable("username") String username,
+                               @ModelAttribute Course course, Principal principal) {
+        if (username.equalsIgnoreCase(principal.getName())) {
+            courseService.deleteCourse(course.getYear(), principal);
+            return "redirect:/user/{username}/dashboard";
+        }
+        return "error/403";
     }
-
 }
